categorization run with cli
